fix(web/teacher): fix update auth page
