+++ web/teacher/src/pages/settings/password.vue
@@ -28,7 +28,7 @@ import { useRouter } from '@nuxtjs/composition-api'
 import { defineComponent, reactive } from '@vue/composition-api'
 import TheFormGroup from '~/components/atoms/TheFormGroup.vue'
 import TheTextField from '~/components/atoms/TheTextField.vue'
-import { CommonStore, UserStore } from '~/store'
+import { AuthStore, CommonStore } from '~/store'
 import { TeacherUpdatePasswordOptions, TeacherUpdatePasswordParams } from '~/types/form'
 
 export default defineComponent({
@@ -47,8 +47,9 @@ export default defineComponent({
     const handleSubmit = async (): Promise<void> => {
       CommonStore.startConnection()
 
-      await UserStore.updatePassword({ form: updatePasswordForm })
+      await AuthStore.updatePassword({ form: updatePasswordForm })
         .then(() => {
+          AuthStore.signOut()
           router.push('/signin')
           CommonStore.showSnackbar({ color: 'success', message: 'パスワードを更新しました。' })
         })
