+++ web/teacher/src/types/api/v1/auth.ts
@@ -1,5 +1,24 @@
 import { SubjectResponse } from '~/types/api/v1'
 
+/**
+ * ---------------------------
+ * Request
+ * ---------------------------
+ */
+export interface UpdateMyMailRequest {
+  mail: string
+}
+
+export interface UpdateMyPasswordRequest {
+  password: string
+  passwordConfirmation: string
+}
+
+/**
+ * ---------------------------
+ * Response
+ * ---------------------------
+ */
 export interface AuthResponse {
   id: string
   lastName: string
